+++ terrain/voxel_lod_terrain.cpp
@@ -44,11 +44,11 @@ Ref<VoxelStream> VoxelLodTerrain::get_stream() const {
 	return _stream;
 }
 
-int VoxelLodTerrain::get_block_size() const {
+unsigned int VoxelLodTerrain::get_block_size() const {
 	return _lods[0].map->get_block_size();
 }
 
-int VoxelLodTerrain::get_block_size_pow2() const {
+unsigned int VoxelLodTerrain::get_block_size_pow2() const {
 	return _lods[0].map->get_block_size_pow2();
 }
 
@@ -70,7 +70,7 @@ void VoxelLodTerrain::set_stream(Ref<VoxelStream> p_stream) {
 }
 
 void VoxelLodTerrain::make_all_view_dirty_deferred() {
-	for (int i = 0; i < get_lod_count(); ++i) {
+	for (unsigned int i = 0; i < get_lod_count(); ++i) {
 		Lod &lod = _lods[i];
 		lod.last_view_distance_blocks = 0;
 	}
@@ -153,11 +153,11 @@ void VoxelLodTerrain::set_lod_count(unsigned int p_lod_count) {
 
 	if (get_lod_count() != p_lod_count) {
 
-		int bs = get_block_size();
+		unsigned int bs = get_block_size();
 		LodOctree<bool>::NoDestroyAction nda;
 		_lod_octree.create_from_lod_count(bs, p_lod_count, nda);
 
-		for (int lod_index = 0; lod_index < MAX_LOD; ++lod_index) {
+		for (unsigned int lod_index = 0; lod_index < MAX_LOD; ++lod_index) {
 
 			Lod &lod = _lods[lod_index];
 
@@ -181,7 +181,7 @@ void VoxelLodTerrain::set_lod_count(unsigned int p_lod_count) {
 	}
 }
 
-int VoxelLodTerrain::get_lod_count() const {
+unsigned int VoxelLodTerrain::get_lod_count() const {
 	return _lod_octree.get_lod_count();
 }
 
@@ -391,7 +391,7 @@ bool VoxelLodTerrain::check_block_loaded_and_updated(const Vector3i &p_bpos, uns
 //		Rect3i box,
 //		Set<Vector3i> &position_set) {
 
-//	for (int i = 0; i < positions.size(); ++i) {
+//	for (unsigned int i = 0; i < positions.size(); ++i) {
 //		const Vector3i bpos = positions[i];
 //		if (!box.contains(bpos)) {
 //			int last = positions.size() - 1;
@@ -434,7 +434,7 @@ void VoxelLodTerrain::_process() {
 			// Each LOD keeps a box of loaded blocks, and only some of the blocks will get polygonized.
 			// The player can edit them so changes can be propagated to lower lods.
 
-			int block_size_po2 = _lods[0].map->get_block_size_pow2() + lod_index;
+			unsigned int block_size_po2 = _lods[0].map->get_block_size_pow2() + lod_index;
 			Vector3i viewer_block_pos_within_lod = VoxelMap::voxel_to_block_b(viewer_pos, block_size_po2);
 
 			Rect3i new_box = Rect3i::from_center_extents(viewer_block_pos_within_lod, Vector3i(block_region_extent));
@@ -483,7 +483,7 @@ void VoxelLodTerrain::_process() {
 	{
 		struct SubdivideAction {
 			VoxelLodTerrain *self;
-			int blocked_count = 0;
+			unsigned int blocked_count = 0;
 
 			bool can_do(LodOctree<bool>::Node *node, unsigned int lod_index) {
 				CRASH_COND(lod_index == 0);
@@ -513,7 +513,7 @@ void VoxelLodTerrain::_process() {
 
 		struct UnsubdivideAction {
 			VoxelLodTerrain *self;
-			int blocked_count = 0;
+			unsigned int blocked_count = 0;
 
 			bool can_do(LodOctree<bool>::Node *node, unsigned int lod_index) {
 				// Can only unsubdivide if the parent mesh is ready
@@ -559,7 +559,7 @@ void VoxelLodTerrain::_process() {
 		for (unsigned int lod_index = 0; lod_index < get_lod_count(); ++lod_index) {
 			Lod &lod = _lods[lod_index];
 
-			for (int i = 0; i < lod.blocks_to_load.size(); ++i) {
+			for (unsigned int i = 0; i < lod.blocks_to_load.size(); ++i) {
 				VoxelDataLoader::InputBlock input_block;
 				input_block.position = lod.blocks_to_load[i];
 				input_block.lod = lod_index;
@@ -585,7 +585,7 @@ void VoxelLodTerrain::_process() {
 
 		//print_line(String("Loaded {0} blocks").format(varray(output.emerged_blocks.size())));
 
-		for (int i = 0; i < output.blocks.size(); ++i) {
+		for (unsigned int i = 0; i < (unsigned int)output.blocks.size(); ++i) {
 
 			const VoxelDataLoader::OutputBlock &ob = output.blocks[i];
 
@@ -647,7 +647,7 @@ void VoxelLodTerrain::_process() {
 		for (unsigned int lod_index = 0; lod_index < get_lod_count(); ++lod_index) {
 			Lod &lod = _lods[lod_index];
 
-			for (int i = 0; i < lod.blocks_pending_update.size(); ++i) {
+			for (unsigned int i = 0; i < lod.blocks_pending_update.size(); ++i) {
 				Vector3i block_pos = lod.blocks_pending_update[i];
 
 				VoxelBlock *block = lod.map->get_block(block_pos);
@@ -696,7 +696,7 @@ void VoxelLodTerrain::_process() {
 			_block_updater->pop(output);
 			_stats.updater = output.stats;
 
-			for (unsigned int i = 0; i < output.blocks.size(); ++i) {
+			for (unsigned int i = 0; i < (unsigned int)output.blocks.size(); ++i) {
 				const VoxelMeshUpdater::OutputBlock &ob = output.blocks[i];
 
 				if (ob.lod >= get_lod_count()) {
@@ -712,7 +712,7 @@ void VoxelLodTerrain::_process() {
 		Ref<World> world = get_world();
 		uint32_t timeout = os.get_ticks_msec() + MAIN_THREAD_MESHING_BUDGET_MS; // Allocate milliseconds max to upload meshes
 
-		int queue_index = 0;
+		unsigned int queue_index = 0;
 
 		// The following is done on the main thread because Godot doesn't really support multithreaded Mesh allocation.
 		// This also proved to be very slow compared to the meshing process itself...
@@ -752,9 +752,9 @@ void VoxelLodTerrain::_process() {
 			Ref<ArrayMesh> mesh;
 			mesh.instance();
 
-			int surface_index = 0;
+			unsigned int surface_index = 0;
 			const VoxelMeshUpdater::OutputBlockData &data = ob.data;
-			for (int i = 0; i < data.smooth_surfaces.surfaces.size(); ++i) {
+			for (unsigned int i = 0; i < (unsigned int)data.smooth_surfaces.surfaces.size(); ++i) {
 
 				Array surface = data.smooth_surfaces.surfaces[i];
 				if (surface.empty()) {
